// berkeley_AI_hackaton/.gadget/server/src/types.ts
export interface BaseRecord {
  __typename?: string;
  id?: string;
  state?: any;
  stateHistory?: any;
  [key: string]: any;
}
// ...
export interface NotYetTyped {
  [key: string]: any;
}

export interface AnyParams {
  [key: string]: string | number | boolean | object | bigint | null | undefined;
}

export interface ActionDescriptor {
  key: string;
  type: "Action" | "ModelAction";
  apiIdentifier: string;
  timeoutMilliseconds: number;
  hasReturnType: boolean;
}

export interface GlobalActionDescriptor {
  key: string;
  type: "GlobalAction";
  apiIdentifier: string;
  timeoutMilliseconds: number;
  hasReturnType: boolean;
}

export interface ModelDescriptor extends ModelMetadata {
  validator: {
    validate(context: AnyEffectContext, record: GadgetRecord<BaseRecord>): Promise<void>;
  };
}
// ...
export interface BaseActionContext extends AnyAmbientContext {
  /** Details about what triggered this action or global action to run */
  trigger: ActionTrigger;

  /**
   * An object passed between all preconditions and effects of an action execution at the \`scope\` property.
   * Useful for transferring data between effects.
   */
  scope: ActionExecutionScope;

  /**
   * @internal
   */
  [AppTenancyKey]?: AppTenancy;
}
// ...
export interface AnyEffectContext extends BaseActionContext {
  type: "effect";
  effect: {
    spec: { id: string };
    configuration: Record<string, any>;
  };
  /**
   * The parameters passed to this effect.
   */
  params: AnyParams;
  /**
   * The specific connection this effect has been contributed by
   **/
  connection?: unknown;
  model?: ModelDescriptor; // present for effects on actions, not present for effects on GlobalActions
  record?: GadgetRecord<BaseRecord>; // present for effects on actions, not present for effects on GlobalActions
}
// ...
export interface FieldMetadata {
  fieldType: string;
  key: string;
  name: string;
  apiIdentifier: string;
  configuration: { [key: string]: any };
  internalWritable: boolean;
}

/**
 * @hidden
 */
export interface ModelMetadata {
  key: string;
  name: string;
  apiIdentifier: string;
  namespace: string[];
  fields: {
    [key: string]: FieldMetadata;
  };
  graphqlTypeName: string;
  stateChart: any;
}